Improvement html report
Create more complete reports to each type of report that we have.
This change generate a more complete report to Ios and Android report,
since a lot of fields was not showing in final html report, only for
json.
With this change we have different templates for `report.Report`,
`report.IOSReporter` and `report.AndroidReporter` and this templates has
a lot of common desgins duplicated, so in the future, we should
extract this duplicated sessions and use some "default" section of
report.