Fix read symlink files
If a symbolink link exists into target directory, and this symbolink point
to a directory, the check of info.IsDir() returns false, so, when we will
try to read the content of file we got an error because we trying to read
from a directory. To fix this problem, for now, we will ignore symbolink
and on the future we can follow these symbolink and correct read their contents.