Create flag to ignore DRA
This commit add a new flag -no-dra to ignore DRA analysis.
Before this changes, we start a new go routine to each file on target
directory, and this go routine did all the job of load rules, convert,
the file to an InputFile, apply rules and check for DRA.
Now we load the rules and convert to InputFile on the main go routine,
and spawns new go routines to apply the rules and check for DRA (if
enabled). One thing that we can do better on the future is during the
update of Result struct, we have always to Lock and Unlock the mutex,
maybe we should read theses updates from a channel and update the Result
in one place.

Fixes #34